requirements.txt
altair==4.0.0
langchain
openai
streamlit==1.12.0
streamlit-chat==0.0.2.2
pinecone-client